Back/requirements.txt
Flask
Flask-Cors
pymongo[srv]
Flask-PyMongo
requests
beautifulsoup4
python-dotenv
openai
pandas
numpy
transformers
tensorflow
spacy
python-docx
docx2pdf
scikit-learn
convertapi
tf-keras
torch
gunicorn





